microsoft-graph/me/beta/mecalendargroupcalendarcalendarviewexceptionoccurrence/id_mecalendargroupcalendarcalendarview.go
package mecalendargroupcalendarcalendarviewexceptionoccurrence

import (
	"fmt"
	"strings"

	"github.com/hashicorp/go-azure-helpers/resourcemanager/resourceids"
)

// Copyright (c) HashiCorp Inc. All rights reserved.
// Licensed under the MIT License. See NOTICE.txt in the project root for license information.

var _ resourceids.ResourceId = MeCalendarGroupCalendarCalendarViewId{}

// MeCalendarGroupCalendarCalendarViewId is a struct representing the Resource ID for a Me Calendar Group Calendar Calendar View
type MeCalendarGroupCalendarCalendarViewId struct {
	CalendarGroupId string
	CalendarId      string
	EventId         string
}

// NewMeCalendarGroupCalendarCalendarViewID returns a new MeCalendarGroupCalendarCalendarViewId struct
func NewMeCalendarGroupCalendarCalendarViewID(calendarGroupId string, calendarId string, eventId string) MeCalendarGroupCalendarCalendarViewId {
	return MeCalendarGroupCalendarCalendarViewId{
		CalendarGroupId: calendarGroupId,
		CalendarId:      calendarId,
		EventId:         eventId,
	}
}

// ParseMeCalendarGroupCalendarCalendarViewID parses 'input' into a MeCalendarGroupCalendarCalendarViewId
func ParseMeCalendarGroupCalendarCalendarViewID(input string) (*MeCalendarGroupCalendarCalendarViewId, error) {
	parser := resourceids.NewParserFromResourceIdType(MeCalendarGroupCalendarCalendarViewId{})
	parsed, err := parser.Parse(input, false)
	if err != nil {
		return nil, fmt.Errorf("parsing %q: %+v", input, err)
	}

	var ok bool
	id := MeCalendarGroupCalendarCalendarViewId{}

	if id.CalendarGroupId, ok = parsed.Parsed["calendarGroupId"]; !ok {
		return nil, resourceids.NewSegmentNotSpecifiedError(id, "calendarGroupId", *parsed)
	}

	if id.CalendarId, ok = parsed.Parsed["calendarId"]; !ok {
		return nil, resourceids.NewSegmentNotSpecifiedError(id, "calendarId", *parsed)
	}

	if id.EventId, ok = parsed.Parsed["eventId"]; !ok {
		return nil, resourceids.NewSegmentNotSpecifiedError(id, "eventId", *parsed)
	}

	return &id, nil
}

// ParseMeCalendarGroupCalendarCalendarViewIDInsensitively parses 'input' case-insensitively into a MeCalendarGroupCalendarCalendarViewId
// note: this method should only be used for API response data and not user input
func ParseMeCalendarGroupCalendarCalendarViewIDInsensitively(input string) (*MeCalendarGroupCalendarCalendarViewId, error) {
	parser := resourceids.NewParserFromResourceIdType(MeCalendarGroupCalendarCalendarViewId{})
	parsed, err := parser.Parse(input, true)
	if err != nil {
		return nil, fmt.Errorf("parsing %q: %+v", input, err)
	}

	var ok bool
	id := MeCalendarGroupCalendarCalendarViewId{}

	if id.CalendarGroupId, ok = parsed.Parsed["calendarGroupId"]; !ok {
		return nil, resourceids.NewSegmentNotSpecifiedError(id, "calendarGroupId", *parsed)
	}

	if id.CalendarId, ok = parsed.Parsed["calendarId"]; !ok {
		return nil, resourceids.NewSegmentNotSpecifiedError(id, "calendarId", *parsed)
	}

	if id.EventId, ok = parsed.Parsed["eventId"]; !ok {
		return nil, resourceids.NewSegmentNotSpecifiedError(id, "eventId", *parsed)
	}

	return &id, nil
}

// ValidateMeCalendarGroupCalendarCalendarViewID checks that 'input' can be parsed as a Me Calendar Group Calendar Calendar View ID
func ValidateMeCalendarGroupCalendarCalendarViewID(input interface{}, key string) (warnings []string, errors []error) {
	v, ok := input.(string)
	if !ok {
		errors = append(errors, fmt.Errorf("expected %q to be a string", key))
		return
	}

	if _, err := ParseMeCalendarGroupCalendarCalendarViewID(v); err != nil {
		errors = append(errors, err)
	}

	return
}

// ID returns the formatted Me Calendar Group Calendar Calendar View ID
func (id MeCalendarGroupCalendarCalendarViewId) ID() string {
	fmtString := "/me/calendarGroups/%s/calendars/%s/calendarView/%s"
	return fmt.Sprintf(fmtString, id.CalendarGroupId, id.CalendarId, id.EventId)
}

// Segments returns a slice of Resource ID Segments which comprise this Me Calendar Group Calendar Calendar View ID
func (id MeCalendarGroupCalendarCalendarViewId) Segments() []resourceids.Segment {
	return []resourceids.Segment{
		resourceids.StaticSegment("staticMe", "me", "me"),
		resourceids.StaticSegment("staticCalendarGroups", "calendarGroups", "calendarGroups"),
		resourceids.UserSpecifiedSegment("calendarGroupId", "calendarGroupIdValue"),
		resourceids.StaticSegment("staticCalendars", "calendars", "calendars"),
		resourceids.UserSpecifiedSegment("calendarId", "calendarIdValue"),
		resourceids.StaticSegment("staticCalendarView", "calendarView", "calendarView"),
		resourceids.UserSpecifiedSegment("eventId", "eventIdValue"),
	}
}

// String returns a human-readable description of this Me Calendar Group Calendar Calendar View ID
func (id MeCalendarGroupCalendarCalendarViewId) String() string {
	components := []string{
		fmt.Sprintf("Calendar Group: %q", id.CalendarGroupId),
		fmt.Sprintf("Calendar: %q", id.CalendarId),
		fmt.Sprintf("Event: %q", id.EventId),
	}
	return fmt.Sprintf("Me Calendar Group Calendar Calendar View (%s)", strings.Join(components, "\n"))
}
